+++ src/main/java/com/highpass/runspot/session/service/SessionService.java
@@ -3,19 +3,18 @@ package com.highpass.runspot.session.service;
 import com.highpass.runspot.auth.domain.Gender;
 import com.highpass.runspot.auth.domain.User;
 import com.highpass.runspot.auth.repository.UserRepository;
-import com.highpass.runspot.session.domain.AttendanceStatus;
 import com.highpass.runspot.session.domain.GenderPolicy;
 import com.highpass.runspot.session.domain.ParticipationStatus;
 import com.highpass.runspot.session.domain.Session;
 import com.highpass.runspot.session.domain.SessionParticipant;
 import com.highpass.runspot.session.domain.SessionStatus;
-import com.highpass.runspot.session.dto.AttendanceUpdateRequest;
-import com.highpass.runspot.session.dto.SessionCreateRequest;
-import com.highpass.runspot.session.dto.SessionJoinRequest;
-import com.highpass.runspot.session.dto.SessionParticipantResponse;
-import com.highpass.runspot.session.dto.SessionResponse;
-import com.highpass.runspot.session.repository.SessionParticipantRepository;
-import com.highpass.runspot.session.repository.SessionRepository;
+import com.highpass.runspot.session.service.dto.request.AttendanceUpdateRequest;
+import com.highpass.runspot.session.service.dto.request.SessionCreateRequest;
+import com.highpass.runspot.session.service.dto.request.SessionJoinRequest;
+import com.highpass.runspot.session.service.dto.response.SessionParticipantResponse;
+import com.highpass.runspot.session.service.dto.response.SessionResponse;
+import com.highpass.runspot.session.domain.dao.SessionParticipantRepository;
+import com.highpass.runspot.session.domain.dao.SessionRepository;
 import java.util.List;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
